More MSVC fixes

+++ tests/networking_tests/http_examples/http_examples.cpp
@@ -389,7 +389,7 @@ struct HealthStatus
 {
    std::string status;
    std::string version;
-   int uptime_seconds;
+   int64_t uptime_seconds;
    struct
    {
       bool database;
